# Cache assemblies and wasm using content hashes