… src/Components/Blazor/Build/test/BootJsonWriterTest.cs
… src/Components/Blazor/Build/test/GenerateBlazorBootJsonTest.cs
… src/Components/Blazor/testassets/HostedInAspNet.Server/RequestLog.cs
… src/Components/Blazor/testassets/HostedInAspNet.Server/Startup.cs
… src/Components/Web.JS/dist/Release/blazor.server.js
… src/Components/Web.JS/dist/Release/blazor.webassembly.js
… src/Components/Web.JS/src/Boot.WebAssembly.ts
… src/Components/Web.JS/src/Platform/Mono/MonoDebugger.ts
… src/Components/Web.JS/src/Platform/Mono/MonoPlatform.ts
… src/Components/Web.JS/src/Platform/Platform.ts
… src/Components/Web.JS/src/Platform/Url.ts
… src/Components/Web.JS/src/Platform/WebAssemblyResourceLoader.ts
… src/Components/Web.JS/src/Services/NavigationManager.ts
… src/Components/test/E2ETest/Tests/BootResourceCachingTest.cs
… src/Shared/E2ETesting/BrowserFixture.cs