This is a prerequisite for the PWA work, but also benefits people who aren't building a PWA, since it means the client no longer has to make `If-Modified-Since` requests for all the assemblies and the wasm file. It also means that badly-configured servers (not returning etags) won't result in bad perf.

It's an unusually complicated change because:

 * I had to reorganize a lot of the logic for loading assemblies to factor parts out of `MonoPlatform.ts`
 * I had to add some extra logic for switching between "streaming" and "arraybuffer" wasm compilation modes, since we can no longer inherit that from emscripten (reasons in comments)
 * Some reordering was needed in the MSBuild targets, because we now need to collect info about the wasm files earlier so we can include their content hashes in blazor.boot.json
 * The caching logic has to be consumable in two different modes:
    * Evaluating the whole network response and validating the hash before completing the promise (for loading .NET assemblies/pdbs)
    * Returning the network response before it's fully loaded, and then validating the hash and caching as a separate background task (for loading wasm binaries, because streaming compilation only makes sense if you start it immediately and not after waiting for the response data to fully arrive)

Some not-strictly-required improvements I also made along the way:

 * Changed `GenerateBlazorBootJson` so it uses S.T.J rather than `DataContractJsonSerializer`
 * Fixed what seems to be a bug in the MSBuild targets that was double-writing `pdb` files to the output, and was unnecessary logic

The build will fail because I haven't yet updated the `.js` binaries and I think some tests will need to be updated too.

As for testing this in general, I plan to write CTI tests. I considered trying to write E2E tests but:

 * It would involve getting Selenium to report on the network traffic so we could observe whether things were retrieved from cache, which is not something we do elsewhere. It may be achievable, but:
 * Concurrent test execution would break things, because the cache is shared across all browser instances and tabs. To deal with this, we'd have to launch the browser on separate user profile directories for each test, but:
    * That's whole new infrastructure I don't want to build
    * Even if we did create that, it's not likely to be portable across browsers, e.g., when these tests run against actual devices. And it would make for slow tests, creating new user profiles for each.

However I think it will be possible to write pretty comprehensive CTI tests as the tester will be able to check the browser cache contents using dev tools, and verify things look right as far as console log output etc.

### Simulated stats

According the browser dev tools, here's the effect on the time taken to fetch all the resources needed to load the page. This is based on a warm HTTP cache (so static resources such as `.dlls` are returning 304s either way):

| Scenario  | Without this change | With this change |
| ------------- | ------------- | ------------- |
| Slow 3G  | 13.5s  | 4.1s |
| Fast 3G  | 5.1s | 2.4s |
| Localhost (artificial scenario) | 0.6s | 0.1s |

The difference is simply due to eliminating 30+ HTTP requests that all would have returned 304s.

### Real-world stats

I've been trying to come up with an argument that we won't actually need to do this at all, and could rely on doing it all in a service worker instead. Unfortunately the data isn't agreeing with me. I deployed to GitHub pages and accessed it over a reasonable quality 4G connection. After ~20 trials both with and without this feature, measured by "finish" time displayed in browser dev tools:

|   | Without this change | With this change |
| ------------- | ------------- | ------------- |
| 25th centile  | 1.57s | 0.48s |
| Median | 1.68s | 0.71s |
| 75th centile | 2.02s | 0.91s |