@SteveSandersonMS 2020-02-11
> Is there also a way to add some tests for the loader and it's caching?

Please see the very long PR description for details :)