@SteveSandersonMS 2020-02-12
I just realised the manual hash checking is completely unnecessary. We can pass `{ integrity: '<hash>' }` as an option to `fetch` and it's all done for us. This will simplify several aspects of the flow :)