@javiercn 2020-02-13
> I just realised the manual hash checking is completely unnecessary. We can pass `{ integrity: '<hash>' }` as an option to `fetch` and it's all done for us. This will simplify several aspects of the flow :)

I was super excited about this, but I'm sad to be the bearer of bad news :(.

It is not supported in Safari looks like. See [here](https://caniuse.com/#feat=mdn-api_request_integrity)