@SteveSandersonMS 2020-02-13
> It is not supported in Safari looks like.

Good news :) It totally does work, at least in the use case we have here. I verified it on Safari iOS both in the success and failure cases for integrity checking. Caniuse has outdated info for this.

From [Safari's release notes](https://developer.apple.com/safari/technology-preview/release-notes/), it looks like this has been supported since at least release 65, which was ages ago.